# Optimize Enum.fetch/2 for negative and out of bound indexes