@eksperimental 2016-07-06
full output can be found [here](https://gist.github.com/eksperimental/dd5d4fec5988ba51f2191e5a754ed710#file-2_enum_fetch_benchmarking_output-txt)
```
$ mix bench.cmp -d percent
bench/snapshots/2016-07-06_11-04-00.snapshot vs
bench/snapshots/2016-07-06_11-15-30.snapshot

## :"Enum.fetch/2: list, 1000 elements"
out of bound (neg index)    -89.82%
first (neg index)           -89.73%
out of bound                -89.65%
middle (neg index)          -24.77%
last                        -1.84%
middle                      +3.91%
first                       +4.00%
last (neg index)            +201.28%

## :"Enum.fetch/2: list_big, 100000 elements"
out of bound                -90.30%
out of bound (neg index)    -89.84%
first (neg index)           -89.77%
middle (neg index)          -24.19%
last                        +1.25%
middle                      +1.37%
first                       +43.48%
last (neg index)            +257.69%

## :"Enum.fetch/2: list_empty, 0 elements"
out of bound                -69.37%
last (neg index)            -69.00%
out of bound (neg index)    -68.77%
last                        -67.09%
first (neg index)           -20.96%
first                       -10.62%
middle (neg index)          -9.16%
middle                      +1.55%

## :"Enum.fetch/2: list_huge, 1000000 elements"
out of bound                -92.34%
first (neg index)           -92.23%
out of bound (neg index)    -92.14%
middle (neg index)          -53.99%
middle                      -1.90%
last                        +0.08%
first                       +30.74%
last (neg index)            +32.17%

## :"Enum.fetch/2: list_single, 1 elements"
out of bound                -65.43%
out of bound (neg index)    -64.94%
last (neg index)            -56.09%
first (neg index)           -4.39%
first                       -2.52%
middle (neg index)          -1.20%
middle                      -0.54%
last                        +7.12%

## :"Enum.fetch/2: map, 1000 elements"
out of bound (neg index)    -99.82%
out of bound                -99.81%
first (neg index)           -78.33%
middle (neg index)          -46.61%
last (neg index)            -23.41%
last                        -7.20%
first                       +2.47%
middle                      +3.15%

## :"Enum.fetch/2: map_big, 100000 elements"
out of bound (neg index)    -100.00%
out of bound                -100.00%
first (neg index)           -74.42%
middle (neg index)          -47.51%
last (neg index)            -16.67%
middle                      -4.40%
last                        -3.88%
first                       -2.72%

## :"Enum.fetch/2: map_empty, 0 elements"
out of bound (neg index)    -83.00%
out of bound                -82.07%
last (neg index)            -81.24%
last                        -80.58%
middle (neg index)          -74.43%
middle                      -73.55%
first                       -72.88%
first (neg index)           -72.43%

## :"Enum.fetch/2: map_huge, 1000000 elements"
out of bound (neg index)    -100.00%
out of bound                -100.00%
first (neg index)           -75.87%
middle (neg index)          -46.63%
last (neg index)            -15.55%
middle                      -4.78%
first                       -4.50%
last                        -0.14%

## :"Enum.fetch/2: map_single, 1 elements"
out of bound                -86.15%
out of bound (neg index)    -82.28%
last (neg index)            -81.77%
last                        -80.65%
first (neg index)           -75.05%
middle                      -74.26%
middle (neg index)          -72.99%
first                       -72.27%
```