@eksperimental 2016-07-06
not sure if all that information should be included in the commit msg though