@josevalim 2016-07-06
@eksperimental thank you for the PR and the benchmarks.

I don't think we should special case lists and maps with two elements. My advice would be drop the clauses for maps and implement `fetch` for lists altogether (instead of only some cases).

The optimization for negative and out-of-bounds positive indexes are good, although we can make them slightly faster.