@ericmj 2016-07-06
I don't think we should spend too much effort on optimizing `Enum.fetch/2` for maps. I don't see where it would make sense to call `Enum.fetch/2` for maps given that the order is arbitrary.