@eksperimental 2016-07-06
all issues addressed, please guys tell me what you think of it.

please do not merge yet as I would like to update the commit message accordingly, and run benchmark again