@eksperimental 2016-07-10
ready for a final review and to be merged