@eksperimental 2016-07-10
@lexmag code updated after tour comments